Supression de la specification de la longueur du INT de la duration du model Snap pour et re JSLint, tant pis

+++ app/models/Snap.js
@@ -16,7 +16,7 @@ module.exports = function (sequelize, DataTypes) {
             }
         },
         duration: {
-            type: DataTypes.INTEGER(2),
+            type: DataTypes.INTEGER,
             allowNull: false,
             validate: {
                 isInt: {
